+++ changelog.md
@@ -2,13 +2,29 @@
 
 All notable changes to this project will be documented in this file.
 
-## [3.0.1] - 2024-21-01
+## [3.0.2] - 2025-03-13
+In component <DataTable />
 
+### Remove 
+- defaultPaginateNumber - props
+- optionsPaginateNumber - props
+- defaultSelectedKeys - props
+
+### Added 
+- rowsPerPageOptions  - props
+
+### Changed
+- showFilter -> hideFilterSearch - props
+- showHandlePaginate -> hideRowsPerPageOptions - props
+
+## Fixed
+- when rows is new data, remove selection
+
+## [3.0.1] - 2024-01-21
 ### Changed
 - Update Readme with new feature heroui
 
-## [3.0.0] - 2024-20-01
-
+## [3.0.0] - 2024-01-20
 ### Added
 - update version librarys
 
@@ -18,8 +34,7 @@ All notable changes to this project will be documented in this file.
 ## Fixed
 - Options disabled no selected with *all Options* props
 
-## [>=2.1.0] - 2024-13-01
-
+## [>=2.1.0] - 2024-01-03
 ### Added
 - Initial oficial release
 - Support React 18 and 19.0.0